+++ app/views/view.py
@@ -1,150 +1,9 @@
-from PyQt5.QtGui import QFontMetrics
-from PyQt5.QtCore import Qt, pyqtSignal, QTimer, QRect
-from PyQt5.QtWidgets import QMainWindow, QWidget, QVBoxLayout, QHBoxLayout, QLineEdit, QLabel, QPushButton, QListWidget, QListWidgetItem, QProgressBar, QTextEdit, QCheckBox, QPushButton, QTableWidget, QTableWidgetItem, QSizePolicy,QTabWidget, QHeaderView, QStyleOptionButton, QStyle, QSplitter
+from PyQt5.QtCore import Qt, QTimer
+from PyQt5.QtWidgets import QMainWindow, QWidget, QVBoxLayout, QLabel, QListWidgetItem, QTableWidget, QTableWidgetItem, QTabWidget, QHeaderView, QSplitter
 
-
-class CheckableHeaderView(QHeaderView):
-    columns_checked = pyqtSignal(object, list)
-
-    def __init__(self, orientation, parent=None):
-        super().__init__(orientation, parent)
-        self.setSectionsClickable(True)
-        self._checkedSections = set()
-        self.table_id = None
-
-    def paintSection(self, painter, rect, logicalIndex):
-        painter.save()
-        super().paintSection(painter, rect, logicalIndex)
-        painter.restore()
-
-        if logicalIndex in self._checkedSections:
-            state = QStyle.State_On
-        else:
-            state = QStyle.State_Off
-
-        option = QStyleOptionButton()
-        option.rect = QRect(rect.x() + rect.width() // 2 - 6, rect.y() + rect.height() // 2 - 6, 12, 12)
-        option.state = QStyle.State_Enabled | state
-        self.style().drawControl(QStyle.CE_CheckBox, option, painter)
-
-    def mousePressEvent(self, event):
-        index = self.logicalIndexAt(event.pos())
-        if index in self._checkedSections:
-            self._checkedSections.remove(index)
-        else:
-            self._checkedSections.add(index)
-        self.viewport().update()
-        self.columns_checked.emit(self.table_id, list(self._checkedSections))
-        super().mousePressEvent(event)
-        
-    def set_checked_sections(self, checked_sections):
-        self._checkedSections = set(checked_sections)
-        self.viewport().update()
-    
-    def set_all_sections_checked(self):
-        self._checkedSections = set(range(self.count()))
-        self.viewport().update()
-
-
-class UIListItem(QWidget):
-    def __init__(self, data, title):
-        super().__init__()
-        self.data = data
-        self.checkbox = QCheckBox()
-        self.checkbox.setChecked(self.data.checked)
-        self.checkbox.toggled.connect(self.checkbox_toggled)
-
-        label = QLabel(title)
-        label.setSizePolicy(QSizePolicy.Fixed, QSizePolicy.Preferred)
-        layout = QHBoxLayout(self)
-        layout.addWidget(self.checkbox)
-        layout.addWidget(label)
-        layout.addStretch(1)
-        self.setLayout(layout)
-
-    def checkbox_toggled(self):
-        self.data.checked = self.checkbox.isChecked()
-
-    def mousePressEvent(self, event):
-        super().mousePressEvent(event)
-        list_widget = self.parent().parent()
-        list_item = list_widget.itemAt(self.parent().mapToParent(event.pos()))
-        list_widget.setCurrentItem(list_item)
-
-
-class ArticleListItem(UIListItem):
-    def __init__(self, article_data):
-        self.article_id = article_data.pmc_id
-        super().__init__(article_data, article_data.title)
-
-
-class DataListItem(UIListItem):
-    preview_requested = pyqtSignal(object)
-
-    def __init__(self, main_window, file_data, disp_name_func):
-        self.main_window = main_window
-        self.page = main_window.active_elements
-        self.file_id = file_data.id
-        disp_name = disp_name_func(file_data)
-        super().__init__(file_data, disp_name)
-
-    def mousePressEvent(self, event):
-        super().mousePressEvent(event)
-        self.preview_requested.emit(self.file_id)
-
-    def get_disp_name(self, text):
-        font_metrics = QFontMetrics(self.main_window.font())
-        available_width = self.page.supp_files_view.width() - 150
-        return font_metrics.elidedText(text, Qt.ElideMiddle, available_width)
-
-
-class SuppFileListItem(DataListItem):
-    def __init__(self, main_window, file_data):
-        super().__init__(main_window, file_data, lambda fd: self.get_disp_name(fd.url.split('/')[-1]))
-        self.file_url = file_data.url
-
-
-class ProcessedTableListItem(DataListItem):
-    def __init__(self, main_window, file_data):
-        super().__init__(main_window, file_data, lambda fd: self.get_disp_name(fd.id))
-
-
-class CommonPageElements:
-    def __init__(self, page):
-        self.prog_bar = QProgressBar(page)
-        self.prog_bar.setRange(0, 100)
-        self.prog_bar.setValue(0)
-        self.prog_bar.hide()
-        self.article_list = QListWidget(page)
-        self.title_disp = QTextEdit(page)
-        self.title_disp.setPlaceholderText("Title will be shown here")
-        self.abstract_disp = QTextEdit(page)
-        self.abstract_disp.setPlaceholderText("Abstract will be shown here")
-        self.supp_files_view = QListWidget(page)
-        self.previews = QWidget(page)
-        self.previews_layout = QVBoxLayout(self.previews)
-        self.previews.setLayout(self.previews_layout)
-        self.loading_label = QLabel(page)
-        self.loading_label.setAlignment(Qt.AlignCenter)
-
-class SearchPageElements(CommonPageElements):
-    def __init__(self, page):
-        super().__init__(page)
-        self.search_status = QLabel(page)
-        self.query_field = QLineEdit(page)
-        self.search_btn = QPushButton("Search", page)
-        self.stop_search_btn = QPushButton("Stop Search", page)
-        self.stop_search_btn.hide()
-        self.stop_search_btn.setEnabled(False)
-        self.proceed_btn = QPushButton("Proceed", page)
-        
-
-class ProcessedPageElements(CommonPageElements):
-    def __init__(self, page):
-        super().__init__(page)
-        self.query_filter_field = QLineEdit(page)
-        self.filter_btn = QPushButton("Filter", page)
-        self.prune_btn = QPushButton("Prune Tables and Columns", page)        
+from views.custom_components import CheckableHeaderView
+from views.list import ArticleListItem, SuppFileListItem, ProcessedTableListItem
+from views.page import SearchPageElements, ProcessedPageElements
 
 class View(QMainWindow):
     def __init__(self):
